Add MOI.attribute_value_type for attributes.

+++ src/Bridges/Constraint/functionize.jl
@@ -55,7 +55,7 @@ function MOI.get(
     ::ScalarFunctionizeBridge{T,S},
     ::MOI.NumberOfConstraints{MOI.ScalarAffineFunction{T},S},
 ) where {T,S}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
@@ -152,7 +152,7 @@ function MOI.get(
     ::VectorFunctionizeBridge{T,S},
     ::MOI.NumberOfConstraints{MOI.VectorAffineFunction{T},S},
 ) where {T,S}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
