+++ src/Bridges/Constraint/geomean.jl
@@ -168,7 +168,7 @@ function concrete_bridge_type(
 end
 
 # Attributes, Bridge acting as a model
-MOI.get(b::GeoMeanBridge, ::MOI.NumberOfVariables) = length(b.xij)
+MOI.get(b::GeoMeanBridge, ::MOI.NumberOfVariables) = Int64(length(b.xij))
 
 MOI.get(b::GeoMeanBridge, ::MOI.ListOfVariableIndices) = copy(b.xij)
 
@@ -176,21 +176,21 @@ function MOI.get(
     ::GeoMeanBridge{T,F},
     ::MOI.NumberOfConstraints{F,MOI.LessThan{T}},
 ) where {T,F}
-    return 1 # t ≤ x_{l1}/sqrt(N)
+    return Int64(1) # t ≤ x_{l1}/sqrt(N)
 end
 
 function MOI.get(
     b::GeoMeanBridge{T,F,G},
     ::MOI.NumberOfConstraints{G,MOI.RotatedSecondOrderCone},
 ) where {T,F,G}
-    return length(b.socrc)
+    return Int64(length(b.socrc))
 end
 
 function MOI.get(
     b::GeoMeanBridge{T,F,G},
     ::MOI.NumberOfConstraints{G,MOI.Nonnegatives},
 ) where {T,F,G}
-    return (b.d > 2 ? 0 : 1)
+    return Int64(b.d > 2 ? 0 : 1)
 end
 
 function MOI.get(
