+++ src/Bridges/Constraint/norm_to_lp.jl
@@ -143,13 +143,15 @@ function concrete_bridge_type(
 end
 
 # Attributes, Bridge acting as a model
-MOI.get(b::NormOneBridge, ::MOI.NumberOfVariables) = length(b.y)
+MOI.get(b::NormOneBridge, ::MOI.NumberOfVariables) = Int64(length(b.y))
+
 MOI.get(b::NormOneBridge, ::MOI.ListOfVariableIndices) = copy(b.y)
+
 function MOI.get(
     b::NormOneBridge{T,F},
     ::MOI.NumberOfConstraints{F,MOI.Nonnegatives},
 ) where {T,F}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
