+++ src/Bridges/Constraint/relentr_to_exp.jl
@@ -82,7 +82,7 @@ end
 
 # Attributes, Bridge acting as a model
 function MOI.get(bridge::RelativeEntropyBridge, ::MOI.NumberOfVariables)
-    return length(bridge.y)
+    return Int64(length(bridge.y))
 end
 
 function MOI.get(bridge::RelativeEntropyBridge, ::MOI.ListOfVariableIndices)
@@ -93,14 +93,14 @@ function MOI.get(
     bridge::RelativeEntropyBridge{T,F},
     ::MOI.NumberOfConstraints{F,MOI.GreaterThan{T}},
 ) where {T,F}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
     bridge::RelativeEntropyBridge{T,F,G},
     ::MOI.NumberOfConstraints{G,MOI.ExponentialCone},
 ) where {T,F,G}
-    return length(bridge.y)
+    return Int64(length(bridge.y))
 end
 
 function MOI.get(
