+++ src/Bridges/Constraint/scalarize.jl
@@ -59,7 +59,7 @@ function MOI.get(
     bridge::ScalarizeBridge{T,F,S},
     ::MOI.NumberOfConstraints{F,S},
 ) where {T,F,S}
-    return length(bridge.scalar_constraints)
+    return Int64(length(bridge.scalar_constraints))
 end
 
 function MOI.get(
