+++ src/Bridges/Constraint/set_map.jl
@@ -42,7 +42,7 @@ function MOI.get(
     ::SetMapBridge{T,S2,S1,F},
     ::MOI.NumberOfConstraints{F,S2},
 ) where {T,S2,S1,F}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
