+++ src/Bridges/Constraint/slack.jl
@@ -16,14 +16,14 @@ function MOI.get(
     ::AbstractSlackBridge{T,VF,ZS,F},
     ::MOI.NumberOfConstraints{F,ZS},
 ) where {T,VF,ZS,F}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
     ::AbstractSlackBridge{T,VF,ZS,F,S},
     ::MOI.NumberOfConstraints{VF,S},
 ) where {T,VF,ZS,F,S}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
@@ -217,7 +217,7 @@ function concrete_bridge_type(
 end
 
 # Attributes, Bridge acting as a model
-MOI.get(b::ScalarSlackBridge, ::MOI.NumberOfVariables) = 1
+MOI.get(b::ScalarSlackBridge, ::MOI.NumberOfVariables) = Int64(1)
 MOI.get(b::ScalarSlackBridge, ::MOI.ListOfVariableIndices) = [b.slack]
 
 # Attributes, Bridge acting as a constraint
@@ -312,7 +312,8 @@ function concrete_bridge_type(
 end
 
 # Attributes, Bridge acting as a model
-MOI.get(b::VectorSlackBridge, ::MOI.NumberOfVariables) = length(b.slack)
+MOI.get(b::VectorSlackBridge, ::MOI.NumberOfVariables) = Int64(length(b.slack))
+
 MOI.get(b::VectorSlackBridge, ::MOI.ListOfVariableIndices) = copy(b.slack)
 
 # Attributes, Bridge acting as a constraint
