+++ src/Bridges/Constraint/soc_to_nonconvex_quad.jl
@@ -173,7 +173,7 @@ function MOI.get(
     ::AbstractSOCtoNonConvexQuadBridge{T},
     ::MOI.NumberOfConstraints{MOI.ScalarQuadraticFunction{T},MOI.LessThan{T}},
 ) where {T}
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
@@ -190,7 +190,7 @@ function MOI.get(
     bridge::AbstractSOCtoNonConvexQuadBridge{T},
     ::MOI.NumberOfConstraints{MOI.ScalarAffineFunction{T},MOI.GreaterThan{T}},
 ) where {T}
-    return length(bridge.var_pos)
+    return Int64(length(bridge.var_pos))
 end
 
 function MOI.get(
