+++ src/Bridges/Variable/free.jl
@@ -35,7 +35,7 @@ end
 
 # Attributes, Bridge acting as a model
 function MOI.get(bridge::FreeBridge, ::MOI.NumberOfVariables)
-    return length(bridge.variables)
+    return Int64(length(bridge.variables))
 end
 
 function MOI.get(bridge::FreeBridge, ::MOI.ListOfVariableIndices)
@@ -46,7 +46,7 @@ function MOI.get(
     ::FreeBridge,
     ::MOI.NumberOfConstraints{MOI.VectorOfVariables,MOI.Nonnegatives},
 )
-    return 1
+    return Int64(1)
 end
 
 function MOI.get(
